+++ src/core/data/interface.ts
@@ -8,10 +8,12 @@
 
 import {
 
+	Decoders,
 	RequestMethods,
 	RequestQuery,
 	RequestBody,
 	ResponseType,
+	ResponseTypes,
 	CreateRequestOpts,
 	MiddlewareParams
 
@@ -21,13 +23,27 @@ export type MockResponseType =
 	ResponseType |
 	object;
 
+export interface MockCustomResponse {
+	status?: number;
+	responseType?: ResponseTypes;
+	decoders?: Decoders;
+}
+
+export interface MockResponseFunction {
+	(params: MiddlewareParams, response: MockCustomResponse): CanPromise<MockResponseType>;
+}
+
+export type MockResponse =
+	CanPromise<MockResponseType> |
+	MockResponseFunction;
+
 export interface Mock {
 	status?: number;
 	query?: RequestQuery;
 	body?: RequestBody;
 	headers?: Dictionary<CanArray<unknown>>;
 	decoders?: boolean;
-	response: CanPromise<MockResponseType> | ((params: MiddlewareParams) => CanPromise<MockResponseType>);
+	response: MockResponse;
 }
 
 export type Mocks = CanPromise<