+++ src/super/i-block/modules/provide.ts
@@ -17,6 +17,12 @@ export type Classes = Dictionary<
 	true
 >;
 
+export type Styles = Dictionary<
+	string |
+	Array<string> |
+	Dictionary<string>
+>;
+
 export type ClassesCacheNms =
 	'base' |
 	'blocks' |
@@ -112,9 +118,13 @@ export default class Provide {
 	 * Returns an object with base component modifiers
 	 * @param mods - additional modifiers ({modifier: {currentValue: value}} || {modifier: value})
 	 */
-	mods(mods?: Dictionary<ModVal | Dictionary<ModVal>>): Readonly<ModsNTable> {
+	mods(mods?: Dictionary<ModVal | Dictionary<ModVal>>): CanUndef<Readonly<ModsNTable>> {
 		const
 			{baseMods} = this.component;
+
+		if (!baseMods && !mods) {
+			return;
+		}
 
 		const
 			key = JSON.stringify(baseMods) + JSON.stringify(mods),